[fix] 알림 읽음 여부 API 경로 수정
